# api/openapi.yaml
openapi: 3.1.0
info:
  title: AGI Egg Autonomous Intelligent Router API
  version: 2024-09-29
  description: |
    API surface for the AGI Egg Intelligent Router platform, covering streaming
    session management, ingest endpoints, automated manifest dispatch, Intelligent
    Model Selector (IMS) decisions, and the AGI Operations Layer (AOL) autonomy flows.
servers:
  - url: https://router.api.agi-egg.example
    description: Production
  - url: https://staging.router.api.agi-egg.example
    description: Staging
    variables: {}
  - url: http://localhost:8080
    description: Local development
...
paths:
  /v1/sessions:
    post:
      tags: [Sessions]
      summary: Create a streaming session
      description: Provision a scoped streaming session and return ingest/stream endpoints plus autonomy defaults.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreateResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session state
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session metadata including latest autonomy status and stream statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionState'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Sessions]
      summary: Terminate a session
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session terminated
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Unique session identifier returned by POST /v1/sessions
      schema:
        type: string
        pattern: "^sess_[A-Za-z0-9]+$"
...
  schemas:
    AutonomyMode:
      type: string
      enum: [observe, propose, act_scoped, act_global]
    SessionScope:
      type: object
      required: [allowedDomains, maxContextTokens, maxCostUSD, piiPolicy, lang]
      properties:
        allowedDomains:
          type: array
          items:
            type: string
        maxContextTokens:
          type: integer
          minimum: 128
        maxCostUSD:
          type: number
          minimum: 0
        piiPolicy:
          type: string
          enum: [mask, redact, reject]
        lang:
          type: string
        autonomyDefaultMode:
          $ref: '#/components/schemas/AutonomyMode'
        autonomyGuardrails:
          type: object
          description: Additional constraints applied to AOL autonomy
          properties:
            maxActScopes:
              type: array
              items:
                type: string
            requireApprovalFor:
              type: array
              items:
                type: string
    ClientInfo:
      type: object
      required: [ua, origin]
      properties:
        ua:
          type: string
        origin:
          type: string
        actorId:
          type: string
        roles:
          type: array
          items:
            type: string
    SessionCreateRequest:
      type: object
      required: [tenantId, projectId, scope, client]
      properties:
        tenantId:
          type: string
        projectId:
          type: string
        scope:
          $ref: '#/components/schemas/SessionScope'
        client:
          $ref: '#/components/schemas/ClientInfo'
        tags:
          type: array
          items:
            type: string
    SessionCreateResponse:
      type: object
      required: [sessionId, ingest, streamUrl, autonomyMode, expiresAt]
      properties:
        sessionId:
          type: string
        ingest:
          type: object
          required: [textUrl, audioUrl]
          properties:
            textUrl:
              type: string
              format: uri
            audioUrl:
              type: string
              format: uri
        streamUrl:
          type: string
          format: uri
        autonomyMode:
          $ref: '#/components/schemas/AutonomyMode'
        expiresAt:
          type: string
          format: date-time
        scope:
          $ref: '#/components/schemas/SessionScope'
    SessionState:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [active, idle, closed]
        autonomyMode:
          $ref: '#/components/schemas/AutonomyMode'
        metrics:
          type: object
          properties:
            streamLatencyP95Ms:
              type: number
            costPerActiveMinute:
              type: number
            imsAccuracy:
              type: number
        lastIntent:
          $ref: '#/components/schemas/IntentUpdate'
        lastModelDecision:
          $ref: '#/components/schemas/ModelDecision'
        activeProposals:
          type: array
          items:
            $ref: '#/components/schemas/AOLProposal'
...
    ModelDecision:
      type: object
      required: [modelId, rationale]
      properties:
        modelId:
          type: string
        rationale:
          type: string
        costClass:
          type: string
          enum: [low, medium, high]
        latencyBudgetMs:
          type: integer
        overrides:
          type: boolean
        fallbackChain:
          type: array
          items:
            type: string
    IntentUpdate:
      type: object
      required: [intent, confidence]
      properties:
        intent:
          type: string
        confidence:
          type: number
        slots:
          type: object
          additionalProperties: true
        selectedModel:
          type: string
        safetyFlags:
          type: array
          items:
            type: string
...
    AOLProposalInput:
      type: object
      required: [summary, scope, playbookId]
      properties:
        summary:
          type: string
        scope:
          type: string
        playbookId:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        autonomyModeRequested:
          $ref: '#/components/schemas/AutonomyMode'
        detector:
          type: object
          properties:
            id:
              type: string
            confidence:
              type: number
            metrics:
              type: object
              additionalProperties:
                type: number
        impactEstimate:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/PlaybookStep'
        verify:
          type: array
          items:
            $ref: '#/components/schemas/VerificationTarget'
    AOLProposal:
      allOf:
        - $ref: '#/components/schemas/AOLProposalInput'
        - type: object
          required: [proposalId, status, createdAt]
          properties:
            proposalId:
              type: string
            status:
              type: string
              enum: [pending, queued, executing, completed, rejected]
            autonomyModeGranted:
              $ref: '#/components/schemas/AutonomyMode'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            execution:
              type: object
              properties:
                startedAt:
                  type: string
                  format: date-time
                completedAt:
                  type: string
                  format: date-time
                outcome:
                  type: string
                  enum: [success, partial, failed]
                metrics:
                  type: object
                  additionalProperties:
                    type: number
    PlaybookStep:
      type: object
      required: [name, tool]
      properties:
        name:
          type: string
        tool:
          type: string
        args:
          type: object
          additionalProperties: true
    VerificationTarget:
      type: object
      required: [metric, success]
      properties:
        metric:
          type: string
        window:
          type: string
        success:
          type: string
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true